requirements.txt
Django==1.9.6
djangorestframework==3.3.3
icalendar==3.10
pip-autoremove==0.9.0
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
six==1.10.0
wheel==0.24.0
